requirements.txt
asgiref==3.5.2
astroid==2.12.12
autopep8==2.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
dill==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
pylint==2.15.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1